Now simply deletes items when there is a match.
Also, there are the proper state changes in game::update between MOVING and RESOLVING based on return value of the relevant function.
However, that checking for matches function is partially implemented (simply deletes 3 or more in a row), but the checking for fall function always returns false - no falling of items occurs.